% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ACE}
\alias{ACE}
\title{ACE (Adverse Childhood Experiences) Dataset}
\format{
A \code{data.frame} with 525 rows and 10 binary (0/1) columns.
}
\source{
Survey of adverse childhood experiences among Chinese students.
}
\usage{
ACE
}
\description{
ACE data from 525 Chinese university students.
  Ten binary items; directly compatible with NIRA.
}
\examples{
data("ACE"); dim(ACE)
}
\keyword{datasets}
