% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GDS9}
\alias{GDS9}
\title{GDS-9 Geriatric Depression Dataset}
\format{
A \code{data.frame} with 3,097 rows and 9 binary (0/1) columns.
}
\source{
Clinical survey of geriatric depression in Chinese elderly.
}
\usage{
GDS9
}
\description{
GDS-9 (Geriatric Depression Scale, 9-item version) data
  from 3,097 Chinese elderly participants. Already binary; no
  transformation required before NIRA.
}
\examples{
data("GDS9"); dim(GDS9)
}
\keyword{datasets}
