% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIRApost-package.R
\docType{package}
\name{NIRApost-package}
\alias{NIRApost}
\alias{NIRApost-package}
\title{NIRApost: Post-Processing Tools for the NodeIdentifyR Algorithm (NIRA)}
\description{
Three-step validation pipeline for NIRA in Ising network analysis:
\enumerate{
  \item Moderation effect testing (prerequisite):
    \code{\link{runMgmmAnalysis}}
  \item Permutation significance testing:
    \code{\link{permutationTest}}, \code{\link{permutationNIRAtest}},
    \code{\link{plotMeanNIRA}}
  \item Repeated-simulation stability assessment:
    \code{\link{stabilityNIRAtest}}, \code{\link{findMaxN}},
    \code{\link{plotMaxN}}
}
}
\section{Data preprocessing utilities}{

\code{\link{checkMissing}}, \code{\link{imputeData}}
}

\section{Datasets}{

\code{single_gds} (GAD-7, N=2404), \code{PHQ9} (N=45829),
\code{GDS9} (N=3097), \code{ACE} (N=525), \code{disease} (N=27353)
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kingfly51/NIRA_post}
  \item Report bugs at \url{https://github.com/kingfly51/NIRA_post/issues}
}

}
\author{
Fei Wang \email{bjnwangfei0501@outlook.com},
  Yiming Wu, Yibo Wu, Tingshao Zhu
}
\keyword{internal}
