% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PHQ9}
\alias{PHQ9}
\title{PHQ-9 Depression Dataset}
\format{
A \code{data.frame} with 45,829 rows and 9 binary (0/1) columns.
}
\source{
2023 PBICR database.
}
\usage{
PHQ9
}
\description{
Binary PHQ-9 responses from the 2023 PBICR database
  (N = 45,829 mainland Chinese participants). "Not at all" = 0;
  all other responses = 1.
}
\examples{
data("PHQ9"); dim(PHQ9)
}
\keyword{datasets}
