% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{single_gds}
\alias{single_gds}
\title{GAD-7 Anxiety Dataset}
\format{
A \code{data.frame} with 2,404 rows and 7 binary (0/1) columns.
}
\source{
National survey on anxiety among Chinese adults.
}
\usage{
single_gds
}
\description{
Binary GAD-7 responses from a national survey of Chinese
  adults (N = 2,404). Original 4-point responses dichotomised:
  "never" = 0; all other responses = 1. Exemplar dataset for the
  NIRApost tutorial.
}
\examples{
data("single_gds")
dim(single_gds)
checkMissing(single_gds)
}
\keyword{datasets}
